fix: Issue(escape)
Issue #1148
