+++ src/Fields/TinyMce.php
@@ -28,6 +28,8 @@ class TinyMce extends Textarea
 
     public array $config = [];
 
+    protected bool $unescape = true;
+
     public function getAssets(): array
     {
         $assets = ["vendor/moonshine/libs/tinymce/tinymce.min.js"];
